criada classe VotacaoService e adicionadas consultas sql personalizadas para Voto

+++ src/main/java/com/example/desafiodigital/repositories/VotacaoRepository.java
@@ -1,7 +1,11 @@
 package com.example.desafiodigital.repositories;
 
 import com.example.desafiodigital.domain.Votacao;
+import com.example.desafiodigital.domain.Voto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VotacaoRepository extends JpaRepository<Votacao, Long> {
+import java.util.Optional;
+
+public interface VotacaoRepository extends JpaRepository<Votacao, Integer> {
+
 }
